+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed `ChocolateySoftware` to be class-based DSC Resource.
 - Changed `ChocolateyPackage` to be class-based DSC Resource.
 - Changed `ChocolateySource` to be a class-based DSC Resource.
+- Changed `ChocolateyFeature` to be a class-based DSC Resource.
 
 ### Added